update the output to int

+++ bmi.py
@@ -3,6 +3,6 @@
 height = float(input('What is your height in m?\n'))
 weight = float(input('What is your weight in kg?\n'))
 
-bmi = weight / height ** 2
+bmi = int(weight / height ** 2)
 
 print('Your body mass index is {}'.format(bmi))
